+++ src/CodeWF.Markdown.Lite.Sample/Program.cs
@@ -1,0 +1,20 @@
+using Avalonia;
+
+namespace CodeWF.Markdown.Lite.Sample;
+
+internal sealed class Program
+{
+    [STAThread]
+    public static void Main(string[] args)
+    {
+        BuildAvaloniaApp().StartWithClassicDesktopLifetime(args);
+    }
+
+    public static AppBuilder BuildAvaloniaApp()
+    {
+        return AppBuilder.Configure<App>()
+            .UsePlatformDetect()
+            .WithInterFont()
+            .LogToTrace();
+    }
+}