+++ src/CodeWF.Markdown.Lite.Sample/Themes/SampleTypographyThemeResources.cs
@@ -1,0 +1,9 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace CodeWF.Markdown.Lite.Sample.Themes;
+
+public partial class SampleTypographyThemeResources : ResourceDictionary
+{
+    public SampleTypographyThemeResources() => AvaloniaXamlLoader.Load(this);
+}