+++ src/CodeWF.Markdown.Lite.Themes/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using Avalonia.Metadata;
+
+[assembly: XmlnsPrefix("https://codewf.com", "markdownThemes")]
+[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.Markdown.Lite.Themes")]