+++ src/CodeWF.Markdown.Lite/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+using Avalonia.Metadata;
+
+[assembly: XmlnsPrefix("https://codewf.com", "markdown")]
+[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.Markdown.Lite")]
+[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.Markdown.Lite.Controls")]
+[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.Markdown.Lite.Rendering")]